# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prescription Reader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Medication {
  id: number;
  prescriptionId: number;
  name: string;
  /** @nullable */
  dosage?: string | null;
  /** @nullable */
  frequency?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Prescription {
  id: number;
  createdAt: string;
  status: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  rawAnalysis?: string | null;
  /**
     * Base64-encoded prescription image
     * @nullable
     */
  imageData?: string | null;
  /**
     * MIME type of the stored image
     * @nullable
     */
  imageMimeType?: string | null;
  medications: Medication[];
}

export interface PrescriptionAnalyzeInput {
  /** Base64-encoded image data */
  imageData: string;
  /** Image MIME type (e.g. image/jpeg) */
  mimeType: string;
}

export interface PrescriptionSummary {
  totalScans: number;
  totalMedications: number;
  recentScans: Prescription[];
}