# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prescription Reader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: prescriptions
    description: Prescription scan operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List all past prescription scans
      responses:
        "200":
          description: List of prescription scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"

  /prescriptions/summary:
    get:
      operationId: getPrescriptionSummary
      tags: [prescriptions]
      summary: Get summary stats of all scans
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionSummary"

  /prescriptions/analyze:
    post:
      operationId: analyzePrescription
      tags: [prescriptions]
      summary: Analyze a prescription image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionAnalyzeInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get a specific prescription scan with its medications
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prescription scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePrescription
      tags: [prescriptions]
      summary: Delete a prescription scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Medication:
      type: object
      properties:
        id:
          type: integer
        prescriptionId:
          type: integer
        name:
          type: string
        dosage:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - prescriptionId
        - name

    Prescription:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        status:
          type: string
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        rawAnalysis:
          type: ["string", "null"]
        imageData:
          type: ["string", "null"]
          description: Base64-encoded prescription image
        imageMimeType:
          type: ["string", "null"]
          description: MIME type of the stored image
        medications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
      required:
        - id
        - createdAt
        - status
        - medications

    PrescriptionAnalyzeInput:
      type: object
      properties:
        imageData:
          type: string
          description: Base64-encoded image data
        mimeType:
          type: string
          description: Image MIME type (e.g. image/jpeg)
      required:
        - imageData
        - mimeType

    PrescriptionSummary:
      type: object
      properties:
        totalScans:
          type: integer
        totalMedications:
          type: integer
        recentScans:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
      required:
        - totalScans
        - totalMedications
        - recentScans