# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prescription Reader API
 * OpenAPI spec version: 0.1.0
 */
import type { Medication } from './medication';

export interface Prescription {
  id: number;
  createdAt: string;
  status: string;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  rawAnalysis?: string | null;
  /**
     * Base64-encoded prescription image
     * @nullable
     */
  imageData?: string | null;
  /**
     * MIME type of the stored image
     * @nullable
     */
  imageMimeType?: string | null;
  medications: Medication[];
}